+++ system/src/Grav/Framework/Collection/ArrayCollection.php
@@ -53,6 +53,37 @@ class ArrayCollection extends BaseArrayCollection implements CollectionInterface
     }
 
     /**
+     * Select items from collection.
+     *
+     * Collection is returned in the order of $keys given to the function.
+     *
+     * @param array $keys
+     * @return static
+     */
+    public function select(array $keys)
+    {
+        $list = [];
+        foreach ($keys as $key) {
+            if ($this->containsKey($key)) {
+                $list[$key] = $this->get($key);
+            }
+        }
+
+        return $this->createFrom($list);
+    }
+
+    /**
+     * Un-select items from collection.
+     *
+     * @param array $keys
+     * @return static
+     */
+    public function unselect(array $keys)
+    {
+        return $this->select(array_diff($this->getKeys(), $keys));
+    }
+
+    /**
      * Implementes JsonSerializable interface.
      *
      * @return array