+++ system/src/Grav/Framework/Collection/CollectionInterface.php
@@ -39,4 +39,22 @@ interface CollectionInterface extends Collection, \JsonSerializable
      * @return array
      */
     public function chunk($size);
+
+    /**
+     * Select items from collection.
+     *
+     * Collection is returned in the order of $keys given to the function.
+     *
+     * @param array $keys
+     * @return static
+     */
+    public function select(array $keys);
+
+    /**
+     * Un-select items from collection.
+     *
+     * @param array $keys
+     * @return static
+     */
+    public function unselect(array $keys);
 }