+++ system/src/Grav/Framework/Object/ObjectIndex.php
@@ -190,6 +190,25 @@ abstract class ObjectIndex extends AbstractIndexCollection implements ObjectColl
     }
 
     /**
+     * @param array $ordering
+     * @return ObjectCollectionInterface
+     */
+    public function orderBy(array $ordering)
+    {
+        return $this->__call('orderBy', [$ordering]);
+    }
+
+    /**
+     * @param int $start
+     * @param int|null $limit
+     * @return static
+     */
+    public function limit($start, $limit = null)
+    {
+        return $this->createFrom($this->slice($start, $limit));
+    }
+
+    /**
      * {@inheritDoc}
      */
     public function call($method, array $arguments = [])